requirements.txt
scapy>=2.5.0
psycopg2-binary>=2.9.9
flask>=3.0.0
